+++ pages/Header.tsx
@@ -1,4 +1,15 @@
 import { ConnectButton } from "@rainbow-me/rainbowkit";
+import type { GetServerSideProps } from "next";
+import { getServerSession } from "next-auth";
+import { getAuthOptions } from "./api/auth/[...nextauth]";
+
+export const getServerSideProps: GetServerSideProps = async ({ req, res }) => {
+  return {
+    props: {
+      session: await getServerSession(req, res, getAuthOptions(req)),
+    },
+  };
+};
 
 const Header = () => {
   return (