+++ pages/_app.tsx
@@ -1,32 +1,83 @@
+import "../styles/global.css";
 import "@rainbow-me/rainbowkit/styles.css";
 import type { AppProps } from "next/app";
-import { useWeb3Login } from "./web3login";
-import Header from "./Header";
+import {
+  RainbowKitProvider,
+  getDefaultWallets,
+  connectorsForWallets,
+} from "@rainbow-me/rainbowkit";
+import {
+  argentWallet,
+  trustWallet,
+  ledgerWallet,
+} from "@rainbow-me/rainbowkit/wallets";
+import { configureChains, createClient, WagmiConfig } from "wagmi";
+import { mainnet, polygon, optimism, arbitrum, goerli } from "wagmi/chains";
+import { publicProvider } from "wagmi/providers/public";
+import { SessionProvider } from "next-auth/react";
+import {
+  RainbowKitSiweNextAuthProvider,
+  GetSiweMessageOptions,
+} from "@rainbow-me/rainbowkit-siwe-next-auth";
+import Header from "./header";
 
-function App({ Component, pageProps }: AppProps) {
-  const {
-    authStatus,
-    RainbowKitProvider,
-    RainbowKitAuthenticationProvider,
-    WagmiConfig,
-    dittiAppInfo,
-    authAdapter,
-    wagmiClient,
-  } = useWeb3Login();
+const { chains, provider, webSocketProvider } = configureChains(
+  [
+    mainnet,
+    polygon,
+    optimism,
+    arbitrum,
+    ...(process.env.NEXT_PUBLIC_ENABLE_TESTNETS === "true" ? [goerli] : []),
+  ],
+  [publicProvider()]
+);
 
+const { wallets } = getDefaultWallets({
+  appName: "Ditti CMS",
+  chains,
+});
+
+const appInfo = {
+  appName: "Ditti CMS",
+};
+
+const connectors = connectorsForWallets([
+  ...wallets,
+  {
+    groupName: "Other",
+    wallets: [
+      argentWallet({ chains }),
+      trustWallet({ chains }),
+      ledgerWallet({ chains }),
+    ],
+  },
+]);
+
+const wagmiClient = createClient({
+  autoConnect: true,
+  connectors,
+  provider,
+  webSocketProvider,
+  version: "2",
+});
+
+const getSiweMessageOptions: GetSiweMessageOptions = () => ({
+  statement: "Sign in to Ditti CMS",
+});
+
+export default function App({ Component, pageProps }: AppProps) {
   return (
-    <WagmiConfig client={wagmiClient}>
-      <RainbowKitAuthenticationProvider
-        adapter={authAdapter}
-        status={authStatus}
-      >
-        <RainbowKitProvider appInfo={dittiAppInfo} chains={dittiAppInfo.chains}>
-          <Header />
-          <Component {...pageProps} />
-        </RainbowKitProvider>
-      </RainbowKitAuthenticationProvider>
-    </WagmiConfig>
+    <SessionProvider refetchInterval={0} session={pageProps.session}>
+      <WagmiConfig client={wagmiClient}>
+        <RainbowKitSiweNextAuthProvider
+          getSiweMessageOptions={getSiweMessageOptions}
+        >
+          <RainbowKitProvider appInfo={appInfo} chains={chains}>
+            <Header />
+            <Component {...pageProps} />
+          </RainbowKitProvider>
+        </RainbowKitSiweNextAuthProvider>
+      </WagmiConfig>
+    </SessionProvider>
   );
 }
-
-export default App;