+++ pages/index.tsx
@@ -1,18 +1,8 @@
-import { ConnectButton } from '@rainbow-me/rainbowkit';
-import type { NextPage } from 'next';
+import { ConnectButton } from "@rainbow-me/rainbowkit";
+import type { NextPage } from "next";
 
 const Home: NextPage = () => {
-  return (
-    <div
-      style={{
-        display: 'flex',
-        justifyContent: 'flex-end',
-        padding: 12,
-      }}
-    >
-      <ConnectButton />
-    </div>
-  );
+  return <div>Test Home Here</div>;
 };
 
 export default Home;