+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# with-next-siwe-iron-session
+# with-next-siwe-next-auth
 
 ## 0.0.17
 
@@ -6,6 +6,7 @@
 
 - Updated dependencies [1876ba0]
   - @rainbow-me/rainbowkit@0.11.0
+  - @rainbow-me/rainbowkit-siwe-next-auth@0.1.8
 
 ## 0.0.16
 
@@ -13,6 +14,7 @@
 
 - Updated dependencies [355402b]
   - @rainbow-me/rainbowkit@0.10.0
+  - @rainbow-me/rainbowkit-siwe-next-auth@0.1.7
 
 ## 0.0.15
 
@@ -20,6 +22,7 @@
 
 - Updated dependencies [49f0ec9]
   - @rainbow-me/rainbowkit@0.9.0
+  - @rainbow-me/rainbowkit-siwe-next-auth@0.1.6
 
 ## 0.0.14
 
@@ -27,12 +30,15 @@
 
 - Updated dependencies [a1d6776]
   - @rainbow-me/rainbowkit@0.8.1
+  - @rainbow-me/rainbowkit-siwe-next-auth@0.1.5
 
 ## 0.0.13
 
 ### Patch Changes
 
 - Updated dependencies [6b37050]
+- Updated dependencies [6b37050]
+  - @rainbow-me/rainbowkit-siwe-next-auth@0.1.5
   - @rainbow-me/rainbowkit@0.8.0
 
 ## 0.0.12
@@ -42,6 +48,7 @@
 - Updated dependencies [e36da59]
 - Updated dependencies [0ff4210]
   - @rainbow-me/rainbowkit@0.7.4
+  - @rainbow-me/rainbowkit-siwe-next-auth@0.1.4
 
 ## 0.0.11
 
@@ -49,6 +56,7 @@
 
 - Updated dependencies [5a65178]
   - @rainbow-me/rainbowkit@0.7.3
+  - @rainbow-me/rainbowkit-siwe-next-auth@0.1.4
 
 ## 0.0.10
 
@@ -56,6 +64,7 @@
 
 - Updated dependencies [1de8203]
   - @rainbow-me/rainbowkit@0.7.2
+  - @rainbow-me/rainbowkit-siwe-next-auth@0.1.4
 
 ## 0.0.9
 
@@ -64,6 +73,7 @@
 - Updated dependencies [6b6a73b]
 - Updated dependencies [5ddc813]
   - @rainbow-me/rainbowkit@0.7.1
+  - @rainbow-me/rainbowkit-siwe-next-auth@0.1.4
 
 ## 0.0.8
 
@@ -71,7 +81,9 @@
 
 - Updated dependencies [2e6bb8f]
 - Updated dependencies [2e6bb8f]
+- Updated dependencies [2e6bb8f]
   - @rainbow-me/rainbowkit@0.7.0
+  - @rainbow-me/rainbowkit-siwe-next-auth@0.1.4
 
 ## 0.0.7
 
@@ -80,6 +92,7 @@
 - Updated dependencies [ecd7209]
 - Updated dependencies [248a1cb]
   - @rainbow-me/rainbowkit@0.6.2
+  - @rainbow-me/rainbowkit-siwe-next-auth@0.1.3
 
 ## 0.0.6
 
@@ -88,6 +101,7 @@
 - Updated dependencies [85eb3bd]
 - Updated dependencies [fbf9d82]
   - @rainbow-me/rainbowkit@0.6.1
+  - @rainbow-me/rainbowkit-siwe-next-auth@0.1.3
 
 ## 0.0.5
 
@@ -96,9 +110,11 @@
 - Updated dependencies [c944ddc]
 - Updated dependencies [c944ddc]
 - Updated dependencies [c944ddc]
+- Updated dependencies [62fd332]
 - Updated dependencies [c944ddc]
 - Updated dependencies [c944ddc]
   - @rainbow-me/rainbowkit@0.6.0
+  - @rainbow-me/rainbowkit-siwe-next-auth@0.1.3
 
 ## 0.0.4
 
@@ -106,6 +122,7 @@
 
 - Updated dependencies [52e2ad6]
   - @rainbow-me/rainbowkit@0.5.3
+  - @rainbow-me/rainbowkit-siwe-next-auth@0.1.2
 
 ## 0.0.3
 
@@ -115,6 +132,7 @@
 - Updated dependencies [fcfc13d]
 - Updated dependencies [3f9013f]
   - @rainbow-me/rainbowkit@0.5.2
+  - @rainbow-me/rainbowkit-siwe-next-auth@0.1.2
 
 ## 0.0.2
 
@@ -124,11 +142,14 @@
 - Updated dependencies [4dfe834]
 - Updated dependencies [8060ccd]
   - @rainbow-me/rainbowkit@0.5.1
+  - @rainbow-me/rainbowkit-siwe-next-auth@0.1.1
 
 ## 0.0.1
 
 ### Patch Changes
 
 - Updated dependencies [737a1d6]
+- Updated dependencies [737a1d6]
 - Updated dependencies [488c5a1]
+  - @rainbow-me/rainbowkit-siwe-next-auth@0.1.6
   - @rainbow-me/rainbowkit@0.5.0